add promo carousel to the homepage
this adds a carousel to the homepage for showing promotional items,
similar to the promotions currently shown on the prod OCW homepage.

the promos themselves are just items in hugo, so there is an archetype
to facilitate adding new ones in the future (see README). this PR adds
three, following the three shown in the mockups.

closes #10
pr #20