fix npm start
small change to the `start.sh` to use `run-p` to start webpack and hugo
via a `startp` npm package script.

so `npm start` to start the app should now work in pretty much all
circumstances. if you have a `.env` file it will export all the
environment variables defined therein and then start `hugo` and
`webpack`, if not it will start `hugo` and `webpack` in the same manner
we were before.

pr #33